# docs/api-spec.yaml
openapi: 3.0.3

info:
  title: Game Log
  version: 1.0.0

paths:
  "/scoreboard":
    summary: Get the Scoreboard of the current active game
    get:
      tags:
        - Scoreboard
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
        '404':
          description: No active game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  "/scoreboard/{game-id}":
    summary: Get the Scoreboard of one specific game
    get:
      tags:
        - Scoreboard
      parameters:
        - in: path
          name: game-id
          schema:
            type: string
            format: uuid
            example: ba0e1deb-db59-4f07-a0d8-4def1f484d31
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scoreboard'
        '404':
          description: No Scoreboard for this game found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

components:
  schemas:
    Player:
      type: object
      required:
        - id
        - name
      properties: 
        id: 
          type: string
          format: uuid
          example: ba0e1deb-db59-4f07-a0d8-4def1f484d31
        name:
          type: string
          example: Team Blue
    ScoreboardEntry:
      type: object
      required:
        - player
        - score
      properties:
        player:
          $ref: '#/components/schemas/Player'
        score:
          type: integer
          format: int32
          example: 1234
    Scoreboard:
      type: object
      required: 
        - scoreborad
      properties:
        scoreboard:
          type: array
          items:
            $ref: '#/components/schemas/ScoreboardEntry'
    ScoreboardCategory:
      type: string
      enum:
        - global
    Trophy:
      type: object
      required:
        - id
        - name
      properties:
        id: 
          type: string
          format: uuid
          example: ba0e1deb-db59-4f07-a0d8-4def1f484d31
        name:
          type: string
          example: First Blood
    PlayerTrophyCabinet:
      type: object
      required:
        - trophies
      properties:
        trophies:
          type: array
          items:
            $ref: '#/components/schemas/Trophy'            
    ApiError:
      type: object
      additionalProperties: true
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          example: NotFoundException
        message:
          type: string
          example: Resource not found
        timestamp:
          type: string
          format: date-time
          example: 2021-11-02T18:52:34+0000
      
tags:
  - name: Trophy
    description: Trophy System
  - name: Scoreboard
    description: Scoreboard